Changed 'error' msg and settings node name
Originally, 'error' was a message I can send between nodes.  However,
'error' is a built-in message so I needed to change it to 'errorMessage'

In the settings json file, I changed the list of nodes to a list of
packages.  This is to differentiate a package (zipped file from a
server) from the actual node that will run in the network.