Modified custom libraries
I changed the custom libraries to create Servers / Clients.  They are
now much much easier to integrate into a new node.  It handles all of
the under-the-hood stuff to communicate with the main server (all the
node-ipc stuff).  There are some examples inside the libraries at the
bottom as well.  Uncomment 'main()' and you can run them.

Added the Helpers library separately as well.  If this keeps getting
bigger, might as well have it separated.