Added hooks and a few new functions to Client.js
The hook system works like this.  When Node A calls an API of Node C,
it sends a message to SamCore with a 'samcore.send' message.  The
data packet associated will have the Node sender name, the destined node
receiver name, the api call, and any additional data needed to make the
api call.  This transmission is re-broadcasted by SamCore to the entire
network as 'receiver.apiCall' with the data packet.  This allows Node B
to create an 'addHook' callback that can intercept this message and do
something with it.  Note, it can't block transmission or alter the
message in any way.

This can also be done on the receiving end. If Node C returns the api
call with new data to Node A, Node B can intercept that with a,
'sender.receiver.apiCall'.

One specific example this can be useful is with a Discord node.  It can
detect when a file is uploaded to the cloud and send a message to
discord without having to change any code on any other nodes.  It can
'hook' into another node's process.

To run a test run, there are test scripts at the end of the Client.js
and Server.js files:

  In the first terminal:
     $ node ./src/Server.js

  In the second terminal:
     $ node ./src/Client.js