Added timeout to promises
One of the big issues here was if a node crashed, it would lock up any
other nodes that were in the middle of an API call.  This commit fixes
that issue.

In the Helpers object, there is now a function called _promise that asks
for a timelimit in ms.  Default is set to 10 seconds and it will return
an error message.