Optimized a few things
Refactored a few things with the new Files implementation in Helpers.js.
Also realized we did not have to check each stage in the EditJsonFile
object to make sure a path existed in a json file.  Instead, it returns
undefined if the path doesn't exist.  This changes those checks to just
check for undefined.